+++ oiio-sys/src/ffi_imagecache.cpp
@@ -168,7 +168,7 @@ imagecache_filename_from_handle(ImageCache& imagecache, ImageHandle* handle)
 bool
 imagecache_get_image_info(ImageCache& imagecache, rust::Str filename,
                           int subimage, int miplevel, rust::Str dataname,
-                          TypeDesc datatype, const uint8_t* data)
+                          TypeDesc datatype, uint8_t* data)
 {
     OIIO::ustring c_filename(filename.data(), filename.size());
     OIIO::ustring c_dataname(dataname.data(), dataname.size());
@@ -180,7 +180,7 @@ bool
 imagecache_get_image_info_with_handle(ImageCache& imagecache, ImageHandle* file,
                                       Perthread* thread_info, int subimage,
                                       int miplevel, rust::Str dataname,
-                                      TypeDesc datatype, const uint8_t* data)
+                                      TypeDesc datatype, uint8_t* data)
 {
     OIIO::ustring c_dataname(dataname.data(), dataname.size());
     return imagecache.get_image_info(file, thread_info, subimage, miplevel,